### 2022-10-29 #4892 @editorialbot
[PRE REVIEW]: High-performance neural population dynamics modeling enabled by scalable computational infrastructure
**Submitting author:** @a9p (<a href="http://orcid.org/0000-0002-7187-9169">Aashish Patel</a>)
**Repository:** https://github.com/TNEL-UCSD/autolfads-deploy
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
