+++ paper/paper.bib
@@ -6,7 +6,8 @@
     number={11},
     pages={1500--1509},
     year={2014},
-    publisher={Nature Publishing Group}
+    publisher={Nature Publishing Group},
+    doi={10.1038/nn.3776}
 }
 
 @article{keshtkaran2021large,
@@ -14,7 +15,8 @@
     author={Keshtkaran, Mohammad Reza and Sedler, Andrew R and Chowdhury, Raeed H and Tandon, Raghav and Basrai, Diya and Nguyen, Sarah L and Sohn, Hansem and Jazayeri, Mehrdad and Miller, Lee E and Pandarinath, Chethan},
     journal={Nature Methods, in press},
     year={2021},
-    publisher={Nature Publishing Group}
+    publisher={Nature Publishing Group},
+    doi={10.1101/2021.01.13.426570}
 }
 
 @article{pandarinath2018inferring,
@@ -25,14 +27,16 @@
     number={10},
     pages={805--815},
     year={2018},
-    publisher={Nature Publishing Group}
+    publisher={Nature Publishing Group},
+    doi={10.1038/s41592-018-0109-9}
 }
 
 @article{jaderberg2017population,
     title={Population based training of neural networks},
     author={Jaderberg, Max and Dalibard, Valentin and Osindero, Simon and Czarnecki, Wojciech M and Donahue, Jeff and Razavi, Ali and Vinyals, Oriol and Green, Tim and Dunning, Iain and Simonyan, Karen and others},
     journal={arXiv preprint arXiv:1711.09846},
-    year={2017}
+    year={2017},
+    doi={10.48550/arXiv.1711.09846}
 }
 
 @inproceedings{moritz2018ray,
@@ -40,7 +44,8 @@
     author={Moritz, Philipp and Nishihara, Robert and Wang, Stephanie and Tumanov, Alexey and Liaw, Richard and Liang, Eric and Elibol, Melih and Yang, Zongheng and Paul, William and Jordan, Michael I and others},
     booktitle={13th USENIX Symposium on Operating Systems Design and Implementation (OSDI 18)},
     pages={561--577},
-    year={2018}
+    year={2018},
+    doi={10.48550/arXiv.1712.05889}
 }
 
 @misc{george2020katib,
@@ -49,7 +54,8 @@
     year={2020},
     eprint={2006.02085},
     archivePrefix={arXiv},
-    primaryClass={cs.DC}
+    primaryClass={cs.DC},
+    doi={10.48550/arXiv.2006.02085}
 }
 
 @dataset{churchland2021mc_maze,     
@@ -57,18 +63,19 @@
     author = {Churchland, Mark and Kaufman, Matthew}, 
     year = {2021}, 
     url = {https://dandiarchive.org/dandiset/000128/draft}, 
-    publisher = {DANDI archive}, 
+    publisher = {DANDI archive}
 }
 
 @inproceedings{pei2021neural,
+    title = {Neural Latents Benchmark ‘21: Evaluating latent variable models of neural population activity},
     author = {Pei, Felix and Ye, Joel and Zoltowski, David and Zoltowski, David and Wu, Anqi and Chowdhury, Raeed and Sohn, Hansem and O\textquotesingle Doherty, Joseph and Shenoy, Krishna V and Kaufman, Matthew and Churchland, Mark and Jazayeri, Mehrdad and Miller, Lee and Pillow, Jonathan and Park, Il Memming and Dyer, Eva and Pandarinath, Chethan},
     booktitle = {Proceedings of the Neural Information Processing Systems Track on Datasets and Benchmarks},
     editor = {J. Vanschoren and S. Yeung},
     pages = {},
[…]
