### 2022-10-31 5913457
docs(joss): change senior author equal contribution to attribute and fix affiliations spacing
