### 2022-11-07 #4892 @emdupre
👋 Hi @a9p, and thank you for your submission to JOSS !

If you have any suggestions for potential reviewers, please let me know by listing their names or GitHub handles (without the @ character, so they don't receive a notification) here.
