### 2022-11-16 #4892 @mattgolub
Hi Elizabeth,

I wish I could participate here, but I have just taken on way too much
recently. I wouldn't be able to give this the attention it deserves. So I
must respectfully decline.

Best,
Matt

--
Matthew Golub, PhD
Assistant Professor
Paul G. Allen School for Computer Science & Engineering
University of Washington
homes.cs.washington.edu/~mgolub

On Wed, Nov 9, 2022 at 5:20 PM Elizabeth DuPre ***@***.***>
wrote:

> 👋 Hi @cellistigs <https://github.com/cellistigs>, @dghoshal-lbl
> <https://github.com/dghoshal-lbl>, @mattgolub
> <https://github.com/mattgolub>, @lyprince <https://github.com/lyprince>,
>
> Would you be willing to review *High-performance neural population
> dynamics modeling enabled by scalable computational infrastructure* for JOSS
> (the Journal of Open Source Software) <https://joss.theoj.org/>?
>
> You can learn more about reviewing for JOSS -- including our conflict of
> interest policy -- here
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>.
>
> The review takes place on GitHub and focuses on the software
> <https://github.com/TNEL-UCSD/autolfads-deploy> and this short paper:
>
> https://github.com/openjournals/joss-papers/blob/joss.04892/joss.04892/10.21105.joss.04892.pdf
>
> If you're available to review this work, please let me know and I'll add
> you as a reviewer. Once I have sufficient reviewers I'll open a dedicated
> review issue.
>
> Thank you !
>
