### 2022-12-01 #4892 @emdupre
👋 Hi @ghennequin, @shrinandj, @richford, @rabernat,

Would you be willing to review _High-performance neural population dynamics modeling enabled by scalable computational infrastructure_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

This submission aims to provide container-native access for training AutoLFADS models across unmanaged (using Ray) and managed (using KubeFlow and Kubernetes) compute clusters.

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/TNEL-UCSD/autolfads-deploy) and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.04892/joss.04892/10.21105.joss.04892.pdf

If you're available to review this work, please let me know and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you !
