### 2022-12-19 #5023 @emdupre
👋  Hi @richford and @tachukao !

Thank you again for agreeing to review this submission ! The review will take place in this issue, and you can generate your individual reviewer checklists by asking editorialbot directly with `@\editorialbot generate my checklist`.

⭐ I know that the holiday season is here, so just to flag : We ask that reviewers provide their feedback within six weeks. If you anticipate any difficulties in working with this timeline, please let me know !

Otherwise, in working through the checklist, you're likely to have specific feedback on autolfads-deploy. Whenever possible, please open relevant issues on the [linked software repository](https://github.com/TNEL-UCSD/autolfads-deploy) (and cross-link them with this issue) rather than discussing them here. This helps to make sure that feedback is translated into actionable items to improve the software !

If you aren't sure how to get started, please see the [Reviewing for JOSS guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) -- and, of course, feel free to ping me with any questions !
