### 2023-02-04 #5023 @tachukao
Hi @emdupre, the citations look good to me in the proofs, but I'm a bit confused about how I should be checking the citation syntax in the text? Thanks for your help!
