### 2023-02-04 TNEL-UCSD/autolfads-deploy#4 @tachukao
[Documentation] Minor suggestions to the top README
I have a couple of minor suggestions to the README, which I think can improve the documentation.

### Intrdouction

* I think the [introduction](https://github.com/TNEL-UCSD/autolfads-deploy#introduction) can benefit from a short description about what problem this repo is trying to solve.
* I really like the solution matrix, but I think it's assuming quite a bit of knowledge about what the three workflows (Local Container, Ray, and Kubeflow) are. This was made very clear in the paper with Figure 1 and the corresponding explanations. I think it might be worth putting that figure into the README and add some description about the use case from (initial exploration on local machine, further exploration on a local "unmanaged cluster", to launching on AWS/giant clusters).

### Installation and Usage
* In the [installation section](https://github.com/TNEL-UCSD/autolfads-deploy#installation--usage), I think it would be good to be consistent with above and change "cluster" to "local cluster".

### Others
* I think it would be helpful o add
     * a section pointing to the `examples` directory and
     * another section pointing to the JOSS paper.pdf

ref: [JOSS review](https://github.com/openjournals/joss-reviews/issues/5023#issuecomment-1371201687)
