### 2023-02-04 TNEL-UCSD/autolfads-deploy#5 @tachukao
[Documentation] Minor suggestions for the examples directory
* Currently, the examples directory contains a description of how to launch auto-lfads on the mc_maze dataset with Kubeflow. It would be great to add additional documentation, explaining how to do the same with local cluster/Ray
* If it's not too difficult to clean up, I think it would also be beneficial to add the scripts used to generate Figures 2 and 3 in the examples directory.
* I saw that there used to be a lorenz toy example. Not entirely sure why it was removed. If it's not too much work, I think it would be really helpful to have a toy example like that for testing out the local cluster solution.

ref: [JOSS review](https://github.com/openjournals/joss-reviews/issues/5023#issuecomment-1371201687)
