### 2023-02-04 TNEL-UCSD/autolfads-deploy#6 @tachukao
[Feature request] CLI for generating templates
### Description

* Please correct me if I'm off-base here, but my current understanding of the intended usage of this software is for users to:
    1. git clone this repo
    2. modify the relevant template yaml files/scripts (e.g., `run_pbt.py`) for their problem/dataset
    3. deploy ray or kubeflow according to instructions in the documentation.
* I believe these templates files are going to make it much easier for neuroscientists to apply AutoLFADS at scale to their own datasets
* However, the current usage would also mean that users would have to either
    * work entirely within this repo with multiple versions of essentially the same files, making it challenging to use git to version control their own projects, or
    * manually copy these template files into their own projects

### Proposal

* I think a potential solution to make this process more streamlined/modular is to provide a (pip installable) command line interface that automatically generates the relevant templates/scripts in a user-specified target directory.
* For example, for folks wanting to deploy autolfads with ray, they could do something like:

  ```sh
  autolfads-deploy generate --strategy ray --target [target_dir]
  ````

   This would generate everything under the `ray` folder in `target_dir`. Folks can then modify those files inside `target_dir`.
* I think this would provide a clear API for how to use this software package, as it wasn't entirely clear to me before. Automatic tests can also be written to test this generation process. Otherwise, it seems like automatic testing is missing at the moment?

ref: [Joss review](https://github.com/openjournals/joss-reviews/issues/5023#issuecomment-1371201687)
