### 2023-02-07 #5023 @emdupre
Hi everyone,

Thanks for your updates !

@yarikoptic I appreciate the offer, and I will certainly keep you in mind. Though I'm sure that the authors would appreciate your feedback regardless 😸

> Hi @emdupre, the citations look good to me in the proofs, but I'm a bit confused about how I should be checking the citation syntax in the text? Thanks for your help!

Thanks for your attention on this, @tachukao ! If the citations are rendering appropriately in the proof, then they're in the appropriate syntax.  As an editor, I'll make sure that all of the formatting information in the actual References section is correct once you and @richford can confirm that all of the relevant work is cited !
