### 2023-02-07 TNEL-UCSD/autolfads-deploy#7 @richford
[Documentation] Minor suggestions for paper.md
I have three minor suggestions for the JOSS paper documentation.

- On [this line in the paper](https://github.com/TNEL-UCSD/autolfads-deploy/blob/5913457d032dce63f32d183732adc3adaea06cff/paper/paper.md?plain=1#L83), the phrase "resemblance to peristimulus time histogram PSTHs" was confusing to me (as someone who is not familiar with the initialism. Perhaps you could enclose the initialism in parentheses like "resemblance to peristimulus time histograms (PSTH)"
- Also, the use of "peri-stimulus" vs "peristimulus" is inconsistent in the paper. Perhaps all instances of "peri-stimulus" could be changed to "peristimulus"
- For the reference to Kubeflow in paper.bib, can you retrieve the swid from software heritage in order to have a long term id on the resource. This can be accomplished by changing the `kubeflow` entry to `@softwareversion`. The `author`, `title`, `year`, and `url` keys can stay the same. And then you'd add a `version` key with the swid link. For example, for the latest version, you would use
```
  version = {\href{https://archive.softwareheritage.org/swh:1:dir:9afaedf1ae583012a7d949100fee98e664376b02;origin=https://github.com/kubeflow/kubeflow;visit=swh:1:snp:32b07a4defabcb9748b8682e741dcfa7dcd66b4b;anchor=swh:1:rev:29e74b13975e5f0729fd731b2df3fcd2beee5690}{swh:1:dir:9afaedf1ae583012a7d949100fee98e664376b02}}
```
This will link to the exact version of Kubeflow that you are referencing and I think it's a good fallback option since Kubeflow does not provide you with a DOI.

ref: [JOSS review](https://github.com/openjournals/joss-reviews/issues/5023#issuecomment-1371201687)
