### 2023-02-07 TNEL-UCSD/autolfads-deploy#8 @richford
[Documentation] Add author contributions statement
In paper.md, could you please add an author contributions statement detailing what each author's respective role is in the software. For example, you could categorize author contribution according to the [CRediT taxonomy](http://dx.doi.org/10.1087/20150211) by including a section

```
## Author contributions

We describe contributions to this paper using
the CRediT taxonomy [@credit].
Writing – Original Draft: <<insert appropriate author initials here>>;
Writing – Review & Editing: <<insert appropriate author initials here>>;
Conceptualization and methodology: <<insert appropriate author initials here>>;
Software and data curation: <<insert appropriate author initials here>>;
Validation: <<insert appropriate author initials here>>;
Resources: <<insert appropriate author initials here>>;
Visualization: <<insert appropriate author initials here>>;
Supervision: <<insert appropriate author initials here>>;
Project Administration: <<insert appropriate author initials here>>;
Funding Acquisition: <<insert appropriate author initials here>>;
```

and then adding a `credit` citation to paper.bib:

```
@article{credit,
  author  = {Brand, Amy and Allen, Liz and Altman, Micah and Hlava, Marjorie and Scott, Jo},
  title   = {Beyond authorship: attribution, contribution, collaboration, and credit},
  journal = {Learned Publishing},
  volume  = {28},
  number  = {2},
  pages   = {151-155},
  doi     = {10.1087/20150211},
  url     = {https://onlinelibrary.wiley.com/doi/abs/10.1087/20150211},
  year    = {2015}
}
```
ref: [JOSS review](https://github.com/openjournals/joss-reviews/issues/5023#issuecomment-1371201687)
