### 2023-02-07 TNEL-UCSD/autolfads-deploy#10 @richford
[Documentation] Add contributing guidelines
Can you please add contribution guidelines in a contributing.md file in either the repository's root, docs, or .github folder. The [GitHub docs have a nice section](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) on how to do this along with some links at the bottom to some example guidelines to get you started.

ref: [JOSS review](https://github.com/openjournals/joss-reviews/issues/5023#issuecomment-1371201687)
