### 2023-02-07 TNEL-UCSD/autolfads-deploy#11 @richford
[Testing] Add automated tests
### Description

Please correct me if I'm wrong about this but I could not find any automated testing or unit testing of the AutoLFADS functionality. It would be good to test some of the functionality in this repo using either manual testing that is documented in the repo or automatic testing (preferred) using GitHub actions or another CI workflow.

### Proposal

In addition to unit testing smaller components of AutoLFADS, perhaps you could include a toy dataset in this repository, something small with predictable AutoLFADS outputs. Then the tests could process the toy data and compare against the known solution.

ref: [JOSS review](https://github.com/openjournals/joss-reviews/issues/5023#issuecomment-1371201687)
