### 2023-02-13 #5023 @emdupre
Hi @tachukao and @richford , thank you for your opened issues to date !

I just wanted to confirm that you're each still working through your individual reviewer checklists. If you are instead done with your initial round of comments and waiting on author feedback, please let me know !

@a9p, just to note : you are welcome to begin addressing reviewer comments as they come in, or you can wait until the reviewers have completed their initial reviews.
