### 2023-02-15 814f470
docs: add contributing and code of conduct docs
