### 2023-02-28 #5023 @emdupre
Hi @a9p, I wanted to check in and confirm how the response is going for you. I believe @richford and @tachukao have now completed their original reviews -- though please correct me if that is not the case ! -- so you should be able to proceed with addressing all of the raised points, as I see you have done in https://github.com/TNEL-UCSD/autolfads-deploy/issues/10.

If you have any questions at this point, please don't hesitate to let me know.
