### 2023-03-01 fde0be8
docs(joss): consistency and clarification changes
