### 2023-03-03 c3b191d
JOSS Feedback - Paper text consistency and clarification (#13)
docs(joss): consistency and clarification changes
