+++ paper/paper.bib
@@ -133,13 +133,11 @@
     doi={10.1038/s41593-018-0095-3}
 }
 
-@misc{kubeflow,
-    author = {A. M. Smith and K. Thaney and M. Hahnel},
+@softwareversion{kubeflow,
     title = {Kubeflow: Machine Learning Toolkit for Kubernetes},
     year = {2018},
-    publisher = {GitHub},
-    journal = {GitHub repository},
-    url = {https://github.com/kubeflow/kubeflow}
+    url = {https://github.com/kubeflow/kubeflow},
+    version = {\href{https://archive.softwareheritage.org/swh:1:dir:086e4c66360c96571dccaa8d12645d4316a6b991;origin=https://github.com/kubeflow/kubeflow;visit=swh:1:snp:698e9549e4522b550ae2fea3a204c49e1843e21b;anchor=swh:1:rev:1e9535ec06cae3823c33f29fc7890df9d32fcaf9}{swh:1:dir:086e4c66360c96571dccaa8d12645d4316a6b991}}
 }
 
 @article{vu2018shared,
