### 2023-03-03 39d97d0
docs(joss): add author contributions section
