### 2023-03-03 eb336ff
JOSS Feedback - Readme clarification (#15)
Signed-off-by: Andrew Sedler <34044884+arsedler9@users.noreply.github.com>
Co-authored-by: Andrew Sedler <34044884+arsedler9@users.noreply.github.com>
