### 2023-03-03 b779944
docs(joss): add funding acknowledgements
