+++ examples/README.md
@@ -1,4 +1,10 @@
-# Data with Persistent Volumes
+# Examples
+
+This directory contains example AutoLFADS experiments. Refer to the individual directories for dataset descriptions and usage instructions.
+
+# Useful Tips
+
+## KubeFlow Data Management
 
 If you don't have a storage network solution that can connected to your KubeFlow cluster, you can copy data to a Persistent Volume Claim (PVC) using the below configuration. The top block creates the storage request (change the amount to the required value), and the bottom creates a simple shell container where you can inspect and move files around as necessary.
 
