+++ examples/data/README.md
@@ -1,5 +1,0 @@
-# Sample Datasets
-
-## MC Maze
-
-As part of the NeuralLatents competition, a set of reference datasets can be used to evaluate AutoLFADS. The core (NWB) data and tools can be downloaded from the [nlb_tools repository](https://github.com/neurallatents/nlb_tools#other-resources). To prepare the NWB formatted data, a convenient conversion utility is provided [here](https://github.com/neurallatents/nlb_tools/blob/main/examples/baselines/autolfads/lfads_data_prep.py).
