+++ examples/lorenz/run_lorenz.py
@@ -1,0 +1,110 @@
+"""
+This is the template script for running PBT with lfads_tf2
+modified to be run with the Lorenz dataset.
+Ref: https://github.com/snel-repo/autolfads-tf2/blob/main/example_scripts/run_pbt.py
+"""
+
+import shutil
+from os import path
+
+import ray
+import yaml
+from lfads_tf2.defaults import DEFAULT_CONFIG_DIR
+from lfads_tf2.utils import flatten
+from ray import tune
+from tune_tf2.models import create_trainable_class
+from tune_tf2.pbt.hps import HyperParam
+from tune_tf2.pbt.schedulers import MultiStrategyPBT
+from tune_tf2.pbt.trial_executor import SoftPauseExecutor
+
+# ---------- PBT I/O CONFIGURATION ----------
+# the default configuration file for the LFADS model
+CFG_PATH = path.join(DEFAULT_CONFIG_DIR, "data/lorenz.yaml")
+# the directory to save PBT runs (usually '~/ray_results')
+PBT_HOME = path.expanduser("~/ray_results")
+# the name of this PBT run (run will be stored at {PBT_HOME}/{PBT_NAME})
+RUN_NAME = "autolfads_lorenz"  # the name of the PBT run
+# the dataset to train the PBT model on
+DATA_DIR = ("data")
+DATA_PREFIX = "lorenz"
+
+# ---------- PBT RUN CONFIGURATION ----------
+# whether to use single machine or cluster
+SINGLE_MACHINE = False
+# the number of workers to use - make sure machine can handle all
+NUM_WORKERS = 40
+# the resources to allocate per model
+RESOURCES_PER_TRIAL = {"cpu": 2, "gpu": 0.5}
+# the hyperparameter space to search
+HYPERPARAM_SPACE = {
+    "TRAIN.LR.INIT": HyperParam(
+        1e-5, 5e-3, explore_wt=0.3, enforce_limits=True, init=0.004
+    ),
+    "MODEL.DROPOUT_RATE": HyperParam(
+        0.0, 0.6, explore_wt=0.3, enforce_limits=True, sample_fn="uniform"
+    ),
+    "MODEL.CD_RATE": HyperParam(
+        0.01, 0.7, explore_wt=0.3, enforce_limits=True, init=0.5, sample_fn="uniform"
+    ),
+    "TRAIN.L2.GEN_SCALE": HyperParam(1e-4, 1e-0, explore_wt=0.8),
+    "TRAIN.L2.CON_SCALE": HyperParam(1e-4, 1e-0, explore_wt=0.8),
+    "TRAIN.KL.CO_WEIGHT": HyperParam(1e-6, 1e-4, explore_wt=0.8),
+    "TRAIN.KL.IC_WEIGHT": HyperParam(1e-5, 1e-3, explore_wt=0.8),
+}
+PBT_METRIC = "smth_val_nll_heldin"
+EPOCHS_PER_GENERATION = 25
+# ---------------------------------------------
+
+# setup the data hyperparameters
+dataset_info = {"TRAIN.DATA.DIR": DATA_DIR, "TRAIN.DATA.PREFIX": DATA_PREFIX}
+# setup initialization of search hyperparameters
+init_space = {name: tune.sample_from(hp.init) for name, hp in HYPERPARAM_SPACE.items()}
+# load the configuration as a dictionary and update for this run
+flat_cfg_dict = flatten(yaml.full_load(open(CFG_PATH)))
+flat_cfg_dict.update(dataset_info)
+flat_cfg_dict.update(init_space)
+# Set the number of epochs per generation
+tuneLFADS = create_trainable_class(EPOCHS_PER_GENERATION)
[…]
