### 2023-03-03 862a7fd
JOSS Feedback - Add author contributions section (#14)
