### 2023-03-06 529c3f4
docs: add note for managing experiments
