### 2023-03-06 5403532
docs: add more detail for which files to copy
Signed-off-by: Andrew Sedler <34044884+arsedler9@users.noreply.github.com>
