### 2023-03-06 8966c84
JOSS Feedback - Managing experiments (#18)
Signed-off-by: Andrew Sedler <34044884+arsedler9@users.noreply.github.com>
Co-authored-by: Andrew Sedler <34044884+arsedler9@users.noreply.github.com>
