### 2023-03-07 0430725
Add figure generation scripts per #5
Signed-off-by: SophiaUCSD16 <jih201@ucsd.edu>
