### 2023-03-07 dc4e39d
JOSS Feedback - figure scripts (#19)
Signed-off-by: SophiaUCSD16 <jih201@ucsd.edu>
Co-authored-by: a9p <58503488+a9p@users.noreply.github.com>
