+++ examples/hp_progression.py
@@ -1,0 +1,184 @@
+import argparse
+import logging
+import pandas as pd
+import seaborn as sns
+from matplotlib import colors
+from matplotlib import pyplot as plt
+from mpl_toolkits.axes_grid1.inset_locator import inset_axes
+
+sns.set()  # Set seaborn as default style
+sns.set_theme()
+sns.set_style("ticks")
+
+def generate_cmap():
+    """
+    Generate the color map
+
+    Returns
+    -------
+    np.array
+        int array representing produced color map
+    """
+    scale_num = 3
+
+    # Darker Starter base on RGB Color Code -
+    cmap_color_prefix = sns.dark_palette(
+        (0.28527278, 0.38012942, 0.82346855, 1), n_colors=10 * scale_num
+    )
+    sns.palplot(cmap_color_prefix)
+    cmap_color_prefix = cmap_color_prefix[-6 * scale_num : -5 * scale_num]
+    cmap_color_prefix = list(cmap_color_prefix)
+
+    # # Configure color palette
+    cmap_color = sns.color_palette("coolwarm", n_colors=20 * scale_num)
+    cmap_color = list(cmap_color)
+    cmap_color = cmap_color[: -12 * scale_num]
+    cmap_color.extend([cmap_color[-1]] * 5 * scale_num)
+    cmap_color_prefix.extend(cmap_color)
+    cmap_color = tuple(cmap_color_prefix)
+    sns.palplot(cmap_color)
+    cmap = colors.ListedColormap(cmap_color, name="from_list")
+    return cmap
+
+
+if __name__ == "__main__":
+    """
+    python3 hp_progression.py
+        --model_dir <model output data>
+        --output_dir <>
+    """
+    parser = argparse.ArgumentParser(
+        description="Generate hyperparameter progression figures"
+    )
+    parser.add_argument(
+        "--model_dir",
+        nargs=1,
+        type=str,
+        dest="model_dir",
+        required=True,
+        help="Directory for the trained model",
+    )
+    parser.add_argument(
+        "--output_dir",
+        nargs=1,
+        type=str,
+        dest="output_dir",
+        required=True,
+        help="Directory to save output figures",
+    )
+    parser.add_argument(
+        "--verbose",
+        dest="is_verbose",
+        action="store_true",
+        default=False,
+        help="Display debug messages",
+    )
+    args = parser.parse_args()
+
+    # Construct Logger (stdio)
+    console_handler = logging.StreamHandler()
+    formatter = None
+    if args.is_verbose:
+        console_handler.setLevel(logging.DEBUG)
+        # time - service_name - level - thread name[calling function] - message
+        formatter = logging.Formatter(
+            "%(asctime)s - %(name)s - %(levelname)s - %(funcName)s -  %(message)s"
+        )
+    else:
+        console_handler.setLevel(logging.INFO)
+        # time - service_name - level - message
+        formatter = logging.Formatter(
+            "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+        )
+    console_handler.setFormatter(formatter)
+    # Add the log configurations to the root logger
+    logging.getLogger("").setLevel(logging.DEBUG)
+    logging.getLogger("").addHandler(console_handler)
+    # Get the logger instantiation
+    logger = logging.getLogger("autolfads_hp_figs")
+    logger.debug("Logger successfully initialized")
+    logger.info("Runtime flags: " + str(vars(args)))
+
+    model_dir = args.model_dir[0]
+    output_dir = args.output_dir[0]
+
[…]
