+++ examples/firing_rate_inference.py
@@ -1,0 +1,340 @@
+import argparse
+import logging
+import warnings
+import os
+
+warnings.filterwarnings("ignore")
+warnings.simplefilter("ignore")
+
+import h5py
+import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
+
+plt.rcParams.update({"font.size": 12})
+import seaborn as sns
+
+try:
+    from nlb_tools import make_tensors, nwb_interface
+except ModuleNotFoundError:
+    print("nlb_tools not found." "Make sure that you have installed them.")
+    os.kill(os.getpid(), 9)
+
+PARAMS = {
+    "mc_maze": {
+        "spk_field": "spikes",
+        "hospk_field": "heldout_spikes",
+        "behavior_source": "data",
+        "behavior_field": "hand_vel",
+        "lag": 100,
+        "make_params": {
+            "align_field": "move_onset_time",
+            "align_range": (-250, 450),
+        },
+        "eval_make_params": {
+            "align_field": "move_onset_time",
+            "align_range": (-250, 450),
+        },
+        "fp_len": 200,
+        "psth_params": {
+            "cond_fields": ["trial_type", "trial_version"],
+            "make_params": {
+                "align_field": "move_onset_time",
+                "align_range": (-250, 450),
+            },
+            "kern_sd": 30,
+        },
+    }
+}
+
+def prep_mask(dataset, trial_split):
+    """
+    Converts trial split to boolean array and combines multiple splits if a list is provided
+    
+    Parameters
+    ----------
+    dataset: NWBDataset
+        NWBDataset that trial mask is made for
+    trial_split : {'train', 'val', 'test'}, array-like, or list, optional
+        Trial splits to include in mask.
+
+    Returns
+    -------
+    np.array
+        Boolean array indicating which trials are within
+        provided split(s)
+    """
+    split_to_mask = (
+        lambda x: (dataset.trial_info.split == x) if isinstance(x, str) else x
+    )
+    if isinstance(trial_split, list):
+        trial_mask = np.any([split_to_mask(split) for split in trial_split], axis=0)
+    elif isinstance(trial_split, str):
+        trial_mask = split_to_mask(trial_split)
+    elif isinstance(trial_split, np.ndarray):
+        trial_mask = np.isin(np.arange(len(dataset.trial_info.split)), trial_split)
+    elif isinstance(trial_split, pd.Series):
+        trial_mask = split_to_mask(trial_split)
+    return trial_mask
+
+
+
+if __name__ == "__main__":
+    """
+    Example script command:
+        python3 firing_rate_inference.py
+            --model_dir <model output data>
+            --input_dir <model input data>
+            --dataset_name mc_maze
+            --dataset_dir <DANDI input data>
+            --output_dir <>
+            --ch_list 16 23 48
+            --cond_list 83 96 7 35 56 61
+    """
+    parser = argparse.ArgumentParser(
+        description="Generate firing rate figures"
+    )
+    parser.add_argument(
+        "--model_dir",
+        nargs=1,
+        type=str,
+        dest="model_dir",
+        required=True,
+        help="Directory for the trained model",
+    )
+    parser.add_argument(
+        "--input_dir",
+        nargs=1,
+        type=str,
+        dest="input_dir",
+        required=True,
+        help="Directory for the input data",
+    )
+    parser.add_argument(
+        "--dataset_dir",
+        nargs=1,
[…]
