### 2023-03-07 d4a4436
refactor: update kubeflow and container workflows, add container_run script, update readme for container runtimes (remove docker-compose).
