+++ examples/mc_maze/container_run.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+docker run --rm -it -v $(pwd):/share ucsdtnel/autolfads:$TAG \
+	--data /share/data \
+	--checkpoint /share/container_output \
+	--config-file /share/data/config.yaml
