### 2023-03-07 43185a1
docs: add note about figure scripts in examples readme
