### 2023-03-09 TNEL-UCSD/autolfads-deploy#6 @a9p
Thank you for your feedback! We overhauled the examples and documentation per your noted concerns. We additionally added some clarification around experiment management as an alternative to the above proposal that should provide reduced user burden/barrier to using this repository. And an end-to-end test via ci was introduced in #17. Feel free to reopen this issue if we missed addressing any of your concerns!
