### 2023-03-10 #5023 @emdupre
Thank you for confirming and for your review, @tachukao  !

@a9p, I've finished reviewing the software and paper, and I'm also very happy with the submission. I do have a few editorial requests on the paper itself :

**On figures and tables :**
- On line 51, a reference to Figure 1 is included though its connection to the surrounding text is not explicit. Could you please amend to something like "and on managed or cloud compute, as illustrated in Figure 1." ?
- On line 92, the text is rendering as `tableTable 1`. Could you please remove the word 'table' before the reference ?
- Currently, Table 2 is not referenced in text. Could you please update the text to include this reference or remove the table ?

**On references :**
- Could you please update the `jaderberg2017population`  reference as:
  ```
  @article{jaderberg2017population,
  doi = {10.48550/ARXIV.2201.11941},
  url = {http://arxiv.org/abs/1711.09846},
  author = {Jaderberg, Max and Dalibard, Valentin and Osindero, Simon and Czarnecki, Wojciech M and Donahue, Jeff and Razavi, Ali and Vinyals, Oriol and Green, Tim and Dunning, Iain and Simonyan, Karen and others},
  title = {Population Based Training of Neural Networks},
  year = {2017},
  eprint={1711.09846},
  archivePrefix={arXiv},
  primaryClass={cs.LG}
  }
  ```
- Could you please update the `churchland2021mc_maze` reference to include the version of the dataset used ? Based on the included URL, I assume you used the Version 'draft', but it would be important to note that explicitly.
- The `keshtkaran2021large` reference has `Nature Methods, in press` as the journal ; please update to `Nature Methods`.
- Could you please update the `keshtkaran2019enabling` reference as:
  ```
  @inproceedings{keshtkaran2019enabling,
   author = {Keshtkaran, Mohammad Reza and Pandarinath, Chethan},
   booktitle = {Advances in Neural Information Processing Systems},
   editor = {H. Wallach and H. Larochelle and A. Beygelzimer and F. d\textquotesingle Alch\'{e}-Buc and E. Fox and R. Garnett},
   pages = {},
   publisher = {Curran Associates, Inc.},
   title = {Enabling hyperparameter optimization in sequential autoencoders for spiking neural data},
   url = {https://proceedings.neurips.cc/paper/2019/file/6948bd44c91acd2b54ecdd1b132f10fb-Paper.pdf},
   volume = {32},
   year = {2019}
  }
  ```

**On general formatting:**
- Please add a comma after 'e.g.' on lines 24, 28, 61, 69, 80, and in the Figure 1 caption.
