### 2023-03-10 00949c9
docs(joss): update paper with editor's feedback
