### 2023-03-10 2192d2a
Update paper.bib
Signed-off-by: Andrew Sedler <34044884+arsedler9@users.noreply.github.com>
