+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{cunningham2014dimensionality,
     title = {Dimensionality reduction for large-scale neural recordings},
     author = {Cunningham, John P and Yu, Byron M},
-    journal = {Nature neuroscience},
+    journal = {Nature Neuroscience},
     volume = {17},
     number = {11},
     pages = {1500--1509},
@@ -14,15 +14,19 @@
     title = {A large-scale neural network training framework for generalized estimation of single-trial population dynamics},
     author = {Keshtkaran, Mohammad Reza and Sedler, Andrew R and Chowdhury, Raeed H and Tandon, Raghav and Basrai, Diya and Nguyen, Sarah L and Sohn, Hansem and Jazayeri, Mehrdad and Miller, Lee E and Pandarinath, Chethan},
     journal = {Nature Methods},
-    year = {2021},
-    publisher = {Nature Publishing Group},
-    doi = {10.1101/2021.01.13.426570}
+    volume = {19},
+    number = {12},
+    pages = {1572--1577},
+    year = {2022},
+    publisher = {Springer Nature},
+    doi = {10.1038/s41592-022-01675-0},
+    url = {https://doi.org/10.1038/s41592-022-01675-0},
 }
 
 @article{pandarinath2018inferring,
     title = {Inferring single-trial neural population dynamics using sequential auto-encoders},
     author = {Pandarinath, Chethan and O’Shea, Daniel J and Collins, Jasmine and Jozefowicz, Rafal and Stavisky, Sergey D and Kao, Jonathan C and Trautmann, Eric M and Kaufman, Matthew T and Ryu, Stephen I and Hochberg, Leigh R and others},
-    journal = {Nature methods},
+    journal = {Nature Methods},
     volume = {15},
     number = {10},
     pages = {805--815},
@@ -109,7 +113,7 @@
 @article{yu2021fast,
     title = {Fast deep neural correspondence for tracking and identifying neurons in C. elegans using semi-synthetic training},
     author = {Yu, Xinwei and Creamer, Matthew S and Randi, Francesco and Sharma, Anuj K and Linderman, Scott W and Leifer, Andrew M},
-    journal = {Elife},
+    journal = {eLife},
     volume = {10},
     pages = {e66410},
     year = {2021},
@@ -131,7 +135,7 @@
 @article{golub2018learning,
     title = {Learning by neural reassociation},
     author = {Golub, Matthew D and Sadtler, Patrick T and Oby, Emily R and Quick, Kristin M and Ryu, Stephen I and Tyler-Kabara, Elizabeth C and Batista, Aaron P and Chase, Steven M and Yu, Byron M},
-    journal = {Nature neuroscience},
+    journal = {Nature Neuroscience},
     volume = {21},
     number = {4},
     pages = {607--616},
