### 2023-03-10 5debe39
JOSS Feedback - Paper edits (#20)
Signed-off-by: Andrew Sedler <34044884+arsedler9@users.noreply.github.com>
Co-authored-by: Andrew Sedler <34044884+arsedler9@users.noreply.github.com>
