+++ paper/paper.bib
@@ -1,69 +1,77 @@
 @article{cunningham2014dimensionality,
-    title={Dimensionality reduction for large-scale neural recordings},
-    author={Cunningham, John P and Yu, Byron M},
-    journal={Nature neuroscience},
-    volume={17},
-    number={11},
-    pages={1500--1509},
-    year={2014},
-    publisher={Nature Publishing Group},
-    doi={10.1038/nn.3776}
+    title = {Dimensionality reduction for large-scale neural recordings},
+    author = {Cunningham, John P and Yu, Byron M},
+    journal = {Nature Neuroscience},
+    volume = {17},
+    number = {11},
+    pages = {1500--1509},
+    year = {2014},
+    publisher = {Nature Publishing Group},
+    doi = {10.1038/nn.3776}
 }
 
 @article{keshtkaran2021large,
-    title={A large-scale neural network training framework for generalized estimation of single-trial population dynamics},
-    author={Keshtkaran, Mohammad Reza and Sedler, Andrew R and Chowdhury, Raeed H and Tandon, Raghav and Basrai, Diya and Nguyen, Sarah L and Sohn, Hansem and Jazayeri, Mehrdad and Miller, Lee E and Pandarinath, Chethan},
-    journal={Nature Methods, in press},
-    year={2021},
-    publisher={Nature Publishing Group},
-    doi={10.1101/2021.01.13.426570}
+    title = {A large-scale neural network training framework for generalized estimation of single-trial population dynamics},
+    author = {Keshtkaran, Mohammad Reza and Sedler, Andrew R and Chowdhury, Raeed H and Tandon, Raghav and Basrai, Diya and Nguyen, Sarah L and Sohn, Hansem and Jazayeri, Mehrdad and Miller, Lee E and Pandarinath, Chethan},
+    journal = {Nature Methods},
+    volume = {19},
+    number = {12},
+    pages = {1572--1577},
+    year = {2022},
+    publisher = {Springer Nature},
+    doi = {10.1038/s41592-022-01675-0},
+    url = {https://doi.org/10.1038/s41592-022-01675-0},
 }
 
 @article{pandarinath2018inferring,
-    title={Inferring single-trial neural population dynamics using sequential auto-encoders},
-    author={Pandarinath, Chethan and O’Shea, Daniel J and Collins, Jasmine and Jozefowicz, Rafal and Stavisky, Sergey D and Kao, Jonathan C and Trautmann, Eric M and Kaufman, Matthew T and Ryu, Stephen I and Hochberg, Leigh R and others},
-    journal={Nature methods},
-    volume={15},
-    number={10},
-    pages={805--815},
-    year={2018},
-    publisher={Nature Publishing Group},
-    doi={10.1038/s41592-018-0109-9}
+    title = {Inferring single-trial neural population dynamics using sequential auto-encoders},
+    author = {Pandarinath, Chethan and O’Shea, Daniel J and Collins, Jasmine and Jozefowicz, Rafal and Stavisky, Sergey D and Kao, Jonathan C and Trautmann, Eric M and Kaufman, Matthew T and Ryu, Stephen I and Hochberg, Leigh R and others},
+    journal = {Nature Methods},
+    volume = {15},
+    number = {10},
+    pages = {805--815},
+    year = {2018},
+    publisher = {Nature Publishing Group},
+    doi = {10.1038/s41592-018-0109-9}
 }
 
 @article{jaderberg2017population,
-    title={Population based training of neural networks},
[…]
