+++ .github/workflows/build-test.yml
@@ -1,40 +1,0 @@
-# name: 'Image Build Test'
-
-# on:
-#   workflow_dispatch:
-#   push:
-#     branches:
-#       - master
-#   pull_request:
-#     types:
-#       - review_requested
-
-# permissions:
-#   contents: read
-
-# # This allows a subsequently queued workflow run to interrupt previous runs
-# concurrency:
-#   group: '${{ github.workflow }} @ ${{ github.event.pull_request.head.label || github.head_ref || github.ref }}'
-#   cancel-in-progress: true
-
-# jobs:
-#   build-image:
-#     runs-on: self-hosted
-#     # runs-on: ubuntu-latest
-#     steps:
-#       - name: Checkout
-#         uses: actions/checkout@v3
-#         with:
-#           persist-credentials: false
-
-#       - name: Build cpu image
-#         run: |
-#           docker build -t ucsdtnel/autolfads:ci \
-#             -f ${GITHUB_WORKSPACE}/images/cpu.Dockerfile \
-#             ${GITHUB_WORKSPACE}/images
-
-#       - name: Cleanup
-#         run: |
-#           docker rmi ucsdtnel/autolfads:ci
-
-#       # Do _not_ push images to hub
