### 2023-03-10 #5023 @a9p
Hi @emdupre, thank you for your suggestions - they should all be addressed with https://github.com/TNEL-UCSD/autolfads-deploy/pull/20. The above checklist is also complete with the requested information provided below:
- Release: [TNEL-UCSD/autolfads-deploy v1.1.0](https://github.com/TNEL-UCSD/autolfads-deploy/releases/tag/v1.1.0)
- Zenodo snapshot confirmed with correct metadata
-  v1.1.0 DOI: `10.5281/zenodo.7719505`
