### 2023-03-12 #5023 @emdupre
Thank you @a9p ; this now looks right on my end !

I'm now happy to recommend autolfads-deploy to the EIC team for publication (!), and I just want to add my congratulations on such an impressive effort 🎉✨

Thank you to @tachukao and @richford for your reviews !! Your volunteer efforts are what keep JOSS moving forward.
