### 2023-03-12 #5023 @emdupre
👋  Hi @openjournals/joss-eics  -- I don't think I've seen this error before ! It looks like there's an error in two cross references and one affiliation, but this is only in the JATS validation, not the CrossRef XML validation. Any suggestions on how to debug these errors ?
