### 2023-03-12 7f40466
fix: convert senior author equal contrib symbol to int
