### 2023-03-12 #5023 @a9p
@emdupre https://github.com/TNEL-UCSD/autolfads-deploy/pull/22 should fix this issue! It looks like the root cause was our using a symbol in affiliations: valid for tex/pdf generation, but non-compliant with JATS.
