### 2023-03-21 89fc7eb
JOSS Feedback - Affiliations update (#23)
docs(joss): add country to affiliations
